+++ fakebook/settings.py
@@ -113,7 +113,7 @@ USE_TZ = True
 # Static files (CSS, JavaScript, Images)
 # https://docs.djangoproject.com/en/4.1/howto/static-files/
 STATICFILES_DIRS = [os.path.join(BASE_DIR / 'static')]
-#STATIC_ROOT = [os.path.join(BASE_DIR / 'static')]
+#STATIC_ROOT = os.path.join(BASE_DIR / 'static')
 STATIC_URL = '/static/'
 MEDIA_ROOT = BASE_DIR / 'media'
 MEDIA_URL = '/media/'
