+++ feed/templates/feed/comment_delete.html
@@ -1,8 +1,8 @@
 <!DOCTYPE html>
 {% extends 'feed/base.html' %}
 {% block content %}
-    <button type="button" class="btn btn-danger btn-sm float-end" data-bs-toggle="modal" data-bs-target="#deletecomment"
-                        style="max-width: fit-content">Delete
+    <button type="button" class="btn btn-danger btn-sm" data-bs-toggle="modal" data-bs-target="#deletecomment">
+        <i class="fa-solid fa-trash"></i>
     </button>
     <!-- Modal -->
     <div class="modal fade" id="deletecomment" tabindex="-1" aria-labelledby="deletecommentModalTitle" aria-hidden="true">
