+++ users/templates/users/profile.html
@@ -10,7 +10,7 @@
 {% endif %}
 <div class="container">
     <div class="row mt-3 mx-3">
-        <div class="card shadow-sm col-md-4 col-sm-6 px-3 pt-3 pb-3">
+        <div class="card shadow-sm col-md-4 col-sm-6 px-3 pt-3 pb-3" style="max-height: 205px">
             <div class="media">
                 <div class="d-flex">
                     <img class="rounded-circle account-img" src="{{ u.profile.profile_pic.url }}" width="100" height="100"/>
@@ -65,9 +65,15 @@
                 {% endif %}
                 {% if request.user == u %}
                                            <!-- Button trigger modal -->
-                        <button type="button" class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#exampleModal"
+                        <button type="button" class="btn btn-primary position-relative" data-bs-toggle="modal" data-bs-target="#exampleModal"
                         style="max-width: fit-content">
                           Friend Requests
+                            {% if rec_friend_requests %}
+                                <span class="position-absolute top-0 start-100 translate-middle badge rounded-pill bg-danger">
+                                    {{ rec_friend_requests.count }}
+                                    <span class="visually-hidden">unread messages</span>
+                                </span>
+                            {% endif %}
                         </button>
                         <!-- Modal -->
                         <div class="modal fade" id="exampleModal" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
@@ -139,8 +145,9 @@
                         </div>
                         </div>
                 {% endif %}
-        </div>
-    <ul class="nav nav-pills mt-1 mx-3" id="pills-tab" role="tablist">
+
+    <div class="col-md-4 col-sm-6 px-3 pt-3 pb-3">
+    <ul class="nav justify-content-center nav-pills" id="pills-tab" role="tablist">
       <li class="nav-item" role="presentation">
             <button class="nav-link active" id="pills-home-tab" data-bs-toggle="pill" data-bs-target="#pills-home"
                     type="button" role="tab" aria-controls="pills-home" aria-selected="true">Timeline</button>
@@ -159,7 +166,7 @@
       </li>
     </ul>
     <div class="tab-content mx-4 pt-2" id="pills-tabContent">
-        <div class="tab-pane fade show active" id="pills-home" role="tabpanel" aria-labelledby="pills-home-tab" tabindex="0">
+        <div class="tab-pane fade show active" style="outline: none" id="pills-home" role="tabpanel" aria-labelledby="pills-home-tab" tabindex="0">
             <button type="button" class="btn text-primary fw-bolder" data-bs-toggle="modal" data-bs-target="#post"
                         style="max-width: fit-content">Add a post
             </button>
@@ -191,9 +198,9 @@
                 </div>
               <div class="col-md-5 me-4">
                 {% for post in posts%}
-                    <div class="card card-signin my-3">
+                    <div class="card card-signin my-3" style="width: 500px">
                       <div class="card-body">
-                        <!--<a class="stretched-link" href="{% url 'post-detail' post.pk %}"></a> --->
+                        <!--<a class="stretched-link" href="{% url 'post-detail' post.pk %}"></a> -->
                         {% if not request.user == post.username%}
                             <a href="{{ post.username.profile.get_absolute_url }}"><img src="{{ post.username.profile.profile_pic.url }}" class="rounded-circle" width="30" height="30" alt=""></a>
                             <a class="text-dark" href="{{ post.username.profile.get_absolute_url }}"><b>{{ post.username }}</b></a>
@@ -203,7 +210,16 @@
                         {% endif %}
                         <br><small class="text-muted">Posted on {{ post.date_posted }}</small>
                         <br><br>
-                        <p class="text-dark">{{ post.description }}</p>
+                        <p class="text-dark text-break">{{ post.description }}</p>
+                        <form class="d-share" method="POST" action="{% url 'share-post' post.pk%}" id="{{ post.pk }}">
+                          {% csrf_token %}
+                          <fieldset class=form-group">
+                              {{ share_form | crispy }}
+                          </fieldset>
+                            <div class="form-group float-end">
+                                        <button class="btn btn-sm btn-primary btn-block text-uppercase" type="submit">Share!</button><br>
+                            </div>
+                      </form>
                         {% if post.pic %}
                             <img class="mx-auto d-block" src="{{ post.pic.url }}" alt=" " style="max-block-size: 300px">
                         {% endif %}
@@ -220,11 +236,7 @@
                                  <a type="button" class="btn btn-primary btn-sm" href="{% url 'post-like' %}" id="likebtn{{ post.id }}"> Like</a>
                              {% endif %}
                          </button>
-                      <form class="d-share" method="POST" action="{% url 'share-post' %}" id="{{ post.id }}">
-                            {% csrf_token %}
-                            {{ share_form | crispy }}
-                        </form>
-                          <button class="btn btn-success btn-sm mt-1" onclick="ShareButton({{ post.pk }})">Share Post</button>
+                          <button class="btn btn-success btn-sm mt-1" onclick="ShareButton({{post.pk}})">Share Post</button>
                       </div>
                     </div>
                 {% endfor %}
@@ -247,7 +259,7 @@
                     <a class="btn btn-outline-primary mb-4" href="?page={{ posts.paginator.num_pages }}">Last</a>
                 {% endif %}
         </div>
-        <div class="tab-pane fade" id="pills-profile" role="tabpanel" aria-labelledby="pills-profile-tab" tabindex="0">
+        <div class="tab-pane fade" id="pills-profile" style="outline: none" role="tabpanel" aria-labelledby="pills-profile-tab" tabindex="0">
               <div class="list-group mt-2" style="max-width: 400px">
                 {% if u.profile.name %}
                     <p class="list-group-item"> Name: {{ u.profile.name }}</p>
@@ -266,7 +278,7 @@
                 {% endif %}
             </div>
         </div>
-      <div class="tab-pane fade" id="pills-contact" role="tabpanel" aria-labelledby="pills-contact-tab" tabindex="0">
+      <div class="tab-pane fade" id="pills-contact" style="outline: none" role="tabpanel" aria-labelledby="pills-contact-tab" tabindex="0">
       <div class="mt-2" style="max-inline-size: fit-content">
         <div class="shadow-sm" style="background-color: white">
         {% if friends_list %}
@@ -304,8 +316,10 @@
           {% endif %}
         </div>
     </div>
-    <div class="tab-pane fade" id="pills-photos" role="tabpanel" aria-labelledby="pills-photos-tab" tabindex="0">:)</div>
+    <div class="tab-pane fade" id="pills-photos" style="outline: none" role="tabpanel" aria-labelledby="pills-photos-tab" tabindex="0">:)</div>
 </div>
+</div>
+    </div>
 </div>
 {% endblock content %}
